Int to Double Change
Refactored Calculator to use Double instead of Int

+++ CalculatorApp/Calculator/Program.cs
@@ -2,17 +2,17 @@
 
 using System.Threading.Channels;
 
-float num1 = 0;
-float num2 = 0;
+double num1 = 0;
+double num2 = 0;
 
 Console.WriteLine("Console Calculator in C#\r");
 Console.WriteLine("------------------------\n");
 
 Console.WriteLine("Type in a number, and then press Enter...");
-num1 = Convert.ToInt32(Console.ReadLine());
+num1 = Convert.ToDouble(Console.ReadLine());
 
 Console.WriteLine("Type in another number, and then press Enter...");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 Console.WriteLine("Choose an option from the following list:");
 Console.WriteLine("\ta - Add");
